LPD-24197 - Refactor. Extract utilities from api to their own files
