+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/utils/getAllPicklists.ts
@@ -1,0 +1,34 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {fetch} from 'frontend-js-web';
+
+import openDefaultFailureToast from './openDefaultFailureToast';
+import {IPickList} from './types';
+
+export default async function getAllPicklists(
+	page: number = 1,
+	items: IPickList[] = []
+) {
+	const response = await fetch(
+		`/o/headless-admin-list-type/v1.0/list-type-definitions?pageSize=100&page=${page}`
+	);
+
+	if (!response.ok) {
+		openDefaultFailureToast();
+
+		return [];
+	}
+
+	const responseJSON = await response.json();
+
+	items = [...items, ...responseJSON.items];
+
+	if (responseJSON.lastPage > page) {
+		items = await getAllPicklists(page + 1, items);
+	}
+
+	return items;
+}
