+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/utils/getAllPicklists.d.ts
@@ -3,10 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {FDSViewType} from './FDSViews';
-import {IField, IPickList} from './utils/types';
-export declare function getFields(fdsView: FDSViewType): Promise<IField[]>;
-export declare function getAllPicklists(
+import {IPickList} from './types';
+export default function getAllPicklists(
 	page?: number,
 	items?: IPickList[]
 ): Promise<IPickList[]>;
